requirements.txt
# Core dependencies
numpy
h5py
pillow
fastapi
uvicorn
pydantic-settings
streamlit
requests

# Machine learning and image processing
torch
torchvision

# Data processing
psutil
tqdm

# Database
sqlalchemy

# Optional for accelerated search
faiss-cpu  # Uncomment for CPU-based FAISS
# faiss-gpu  # Uncomment for GPU-based FAISS (requires CUDA)

# HTTP middleware
python-multipart

# Development tools
pytest
black
flake8